# docs/swagger.yaml
basePath: /taskd/api
definitions:
  controllers.AddTemplateResult:
    properties:
      name:
        type: string
    type: object
  controllers.ResponseData:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  dao.ListTasksResult:
    properties:
      list:
        items:
          $ref: '#/definitions/dao.TaskRec'
        type: array
      total:
        type: integer
    type: object
  dao.PoolResource:
    properties:
      id:
        description: 记录ID
        type: integer
      pool_id:
        description: 任务池ID
        type: string
      res_name:
        description: 资源名称
        type: string
      res_num:
        description: 资源数量
        type: string
    type: object
  dao.TaskObjRec:
    properties:
      args:
        description: 该任务的用户参数， JSON格式
        type: string
      callback:
        description: 回调URL
        type: string
      created_by:
        description: 创建者
        type: string
      extra:
        description: 该任务模板的额外信息， JSON格式
        type: string
      name:
        description: 任务名
        type: string
      namespace:
        description: 数据空间名
        type: string
      pool:
        description: 任务池
        type: string
      project:
        description: 项目名
        type: string
      quotas:
        description: 资源配额， JSON格式
        type: string
      tags:
        description: 标记,可影响调度,JSON格式:[key=value]
        type: string
      template:
        description: 任务模板名
        type: string
      timeout:
        description: 各阶段超时时间， JSON格式
        type: string
      updated_by:
        description: 更新者
        type: string
      uuid:
        description: 任务UUID
        type: string
    type: object
  dao.TaskRec:
    properties:
      args:
        description: 该任务的用户参数， JSON格式
        type: string
      callback:
        description: 回调URL
        type: string
      create_time:
        description: 创建时间
        type: string
      created_by:
        description: 创建者
        type: string
      end_log:
        description: 最终日志
        type: string
      end_time:
        description: 结束运行时间
        type: string
      error:
        description: 出错说明
        type: string
      extra:
        description: 该任务模板的额外信息， JSON格式
        type: string
      name:
        description: 任务名
        type: string
      namespace:
        description: 数据空间名
        type: string
      pool:
        description: 任务池
        type: string
      project:
        description: 项目名
        type: string
      quotas:
        description: 资源配额， JSON格式
        type: string
      running_time:
        description: 开始运行时间
        type: string
      start_time:
        description: 启动时间
        type: string
      status:
        description: 任务状态
        type: string
      tags:
        description: 标记,可影响调度,JSON格式:[key=value]
        type: string
      template:
        description: 任务模板名
        type: string
      timeout:
        description: 各阶段超时时间， JSON格式
        type: string
      update_time:
        description: 最后更新时间
        type: string
      updated_by:
        description: 更新者
        type: string
      uuid:
        description: 任务UUID
        type: string
      warning:
        description: 告警信息
        type: string
      yaml_content:
        description: 部署文件内容
        type: string
    type: object
  dao.TemplateRec:
    properties:
      create_time:
        type: string
      engine:
        type: string
      extra:
        type: string
      name:
        type: string
      schema:
        type: string
      title:
        type: string
    type: object
  service.TaskCommitResult:
    properties:
      uuid:
        type: string
    type: object
  service.TaskLogsResult:
    properties:
      entities:
        items:
          $ref: '#/definitions/task.EntityLogs'
        type: array
      status:
        type: string
      uuid:
        type: string
    type: object
  service.TaskPoolArgs:
    properties:
      config:
        description: 任务池配置，供各类型任务引擎使用
        type: string
      description:
        description: 任务池说明，展示任务池的关键信息，供用户了解任务池工作状况
        type: string
      engine:
        description: 采用的任务执行引擎
        type: string
      pool_id:
        description: 任务池ID
        type: string
      resources:
        items:
          $ref: '#/definitions/dao.PoolResource'
        type: array
      running:
        description: 最大并发任务数
        type: integer
      waiting:
        description: 最大排队任务数
        type: integer
    type: object
  service.TaskPoolResult:
    properties:
      pool_id:
        type: string
    type: object
  service.TaskStatusResult:
    properties:
      name:
        type: string
      status:
        type: string
      template:
        type: string
    type: object
  service.TaskTagsResult:
    properties:
      tags:
        additionalProperties:
          type: string
        type: object
    type: object
  task.EntityLogs:
    properties:
      completed:
        type: boolean
      entity:
        type: string
      logs:
        type: string
    type: object
  task.ResourceItem:
    properties:
      allocate:
        description: 分配量
        type: string
      capacity:
        description: 配置量
        type: string
      name:
        description: 资源名称
        type: string
      remain:
        description: 实际残留
        type: string
    type: object
  task.TaskInstanceSummary:
    properties:
      create_time:
        description: 入队时间
        type: string
      created_by:
        description: 用户标记
        type: string
      end_time:
        description: 结束运行的时间
        type: string
      error:
        description: 出错说明
        type: string
      name:
        description: 任务名字
        type: string
      pool:
        description: 在哪个任务池运行
        type: string
      running_time:
        description: 开始正常运行的时间
        type: string
      start_time:
        description: 启动时间
        type: string
      status:
        description: 任务状态
        type: string
      tags:
        description: 标记
        type: string
      uuid:
        description: 任务唯一ID
        type: string
      warning:
        description: 告警信息
        type: string
    type: object
  task.TaskPoolDetail:
    properties:
      config:
        description: 任务池配置
        type: string
      pool_id:
        description: 任务池标识
        type: string
      resources:
        description: 该任务池的资源信息
        items:
          $ref: '#/definitions/task.ResourceItem'
        type: array
      running:
        description: 正在该任务池运行的任务数
        type: integer
      tasks:
        description: 等待该任务池或在其中运行的任务信息
        items:
          $ref: '#/definitions/task.TaskInstanceSummary'
        type: array
      waiting:
        description: 正在等待该任务池的任务数
        type: integer
    type: object
  task.TaskPoolSummary:
    properties:
      config:
        description: 任务池配置
        type: string
      pool_id:
        description: 任务池标识
        type: string
      running:
        description: 正在该任务池运行的任务数
        type: integer
      waiting:
        description: 正在该任务池等待的任务数
        type: integer
    type: object
info:
  contact:
    name: 张钊晋,郑柏春
    url: http://www.sangfor.com.cn
  description: 任务管理
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://www.sangfor.com.cn
  title: 任务管理
  version: 1.0 版本
paths:
  /v1/pools:
    get:
      consumes:
      - application/json
      description: 列举任务池信息，比如资源用量，运行任务概况
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/task.TaskPoolSummary'
            type: array
      summary: 列举任务池
      tags:
      - 任务池
    post:
      consumes:
      - application/json
      description: 添加一个任务池
      parameters:
      - description: 任务池
        in: body
        name: pool
        required: true
        schema:
          $ref: '#/definitions/service.TaskPoolArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TaskPoolResult'
      summary: 添加一个任务池
      tags:
      - 任务池
  /v1/pools/{name}:
    delete:
      consumes:
      - application/json
      description: 删除任务池，同时删除关联的PoolResource
      parameters:
      - description: 任务池名
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ResponseData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ResponseData'
      summary: 删除任务池
      tags:
      - 任务池
    get:
      consumes:
      - application/json
      description: 获取任务池详情
      parameters:
      - description: 任务池名
        in: path
        name: name
        required: true
        type: string
      - description: 是否获取详情
        in: query
        name: verbose
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.TaskPoolDetail'
      summary: 获取任务池详情
      tags:
      - 任务池
    put:
      consumes:
      - application/json
      description: 更新任务池定义
      parameters:
      - description: 任务池ID
        in: path
        name: name
        required: true
        type: string
      - description: 任务池
        in: body
        name: pools
        required: true
        schema:
          $ref: '#/definitions/service.TaskPoolArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TaskPoolResult'
      summary: 更新任务池
      tags:
      - 任务池
  /v1/reload:
    post:
      consumes:
      - application/json
      description: 重新加载任务池相关配置，适用于池节点变更，节点资源变更，池轮询策略变更
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ResponseData'
      summary: 重新加载任务池相关配置
      tags:
      - 任务池
  /v1/tasks:
    get:
      consumes:
      - application/json
      description: 列举任务
      parameters:
      - description: 任务名称
        in: query
        name: name
        type: string
      - description: 命名空间
        in: query
        name: namespace
        type: string
      - description: 任务属主
        in: query
        name: owner
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 待运行资源池
        in: query
        name: pool
        type: string
      - description: 项目名称
        in: query
        name: project
        type: string
      - description: 排序字段
        in: query
        name: sort
        type: string
      - description: 任务状态
        in: query
        name: status
        type: string
      - description: 任务类名称
        in: query
        name: template
        type: string
      - description: UUID
        in: query
        name: uuid
        type: string
      - description: 输出任务实例详情
        in: query
        name: verbose
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dao.ListTasksResult'
      summary: 列举任务
      tags:
      - 任务
    post:
      consumes:
      - application/json
      description: 提交任务
      parameters:
      - description: 任务对象
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/dao.TaskObjRec'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TaskCommitResult'
      summary: 提交任务
      tags:
      - 任务
  /v1/tasks/{uuid}:
    delete:
      consumes:
      - application/json
      description: 停止任务
      parameters:
      - description: 任务UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 停止任务
      tags:
      - 任务
    get:
      consumes:
      - application/json
      description: 获取任务元数据
      parameters:
      - description: 任务UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dao.TaskRec'
      summary: 获取任务元数据
      tags:
      - 任务
  /v1/tasks/{uuid}/logs:
    get:
      consumes:
      - application/json
      description: 获取任务日志
      parameters:
      - description: 任务UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: 任务实体，任务可能启动多个实体(如线程，POD)
        in: query
        name: entity
        type: string
      - in: query
        name: follow
        type: boolean
      - in: query
        name: tail
        type: integer
      - in: query
        name: timestamps
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TaskLogsResult'
      summary: 获取任务日志
      tags:
      - 任务
  /v1/tasks/{uuid}/status:
    get:
      consumes:
      - application/json
      description: 获取任务状态
      parameters:
      - description: 任务UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TaskStatusResult'
      summary: 获取任务状态
      tags:
      - 任务
  /v1/tasks/{uuid}/tags:
    get:
      consumes:
      - application/json
      description: 获取任务标记
      parameters:
      - description: 任务UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TaskTagsResult'
      summary: 获取任务标记
      tags:
      - 任务
    post:
      consumes:
      - application/json
      description: 给任务打标记，通常用于通知任务调度器对任务执行特定处置策略，比如标记为保障任务，空闲任务
      parameters:
      - description: 任务UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: 标记内容，格式为Key=Value,可以同时有多组标记
        in: body
        name: tags
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 给任务打标记
      tags:
      - 任务
  /v1/templates:
    get:
      consumes:
      - application/json
      description: 任务模板
      parameters:
      - description: 是否获取详情
        in: query
        name: verbose
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dao.TemplateRec'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ResponseData'
      summary: 列举任务模板
      tags:
      - 任务模板
    post:
      consumes:
      - application/json
      description: 创建任务模板
      parameters:
      - description: 任务模板
        in: body
        name: templates
        required: true
        schema:
          $ref: '#/definitions/dao.TemplateRec'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.AddTemplateResult'
      summary: 创建任务模板
      tags:
      - 任务模板
  /v1/templates/{name}:
    delete:
      consumes:
      - application/json
      description: 删除任务模板
      parameters:
      - description: 任务模板名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ResponseData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ResponseData'
      summary: 删除任务模板
      tags:
      - 任务模板
    get:
      consumes:
      - application/json
      description: 获取任务模板
      parameters:
      - description: 任务模板名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dao.TemplateRec'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ResponseData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ResponseData'
      summary: 获取任务模板
      tags:
      - 任务模板
    put:
      consumes:
      - application/json
      description: 更新任务模板
      parameters:
      - description: 任务模板名称
        in: path
        name: name
        required: true
        type: string
      - description: 任务模板
        in: body
        name: templates
        required: true
        schema:
          $ref: '#/definitions/dao.TemplateRec'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.AddTemplateResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ResponseData'
      summary: 更新任务模板
      tags:
      - 任务模板
swagger: "2.0"